StorageScan gRPC service implementation, rename files for clarity
