+++ src/control/server/ctl_svc.go
@@ -33,7 +33,6 @@ import (
 
 	"github.com/daos-stack/daos/src/control/common"
 	pb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
-	types "github.com/daos-stack/daos/src/control/common/storage"
 	"github.com/daos-stack/daos/src/control/drpc"
 	log "github.com/daos-stack/daos/src/control/logging"
 )
@@ -79,132 +78,6 @@ func (c *ControlService) Teardown() {
 		log.Debugf(
 			"%s\n", errors.Wrap(err, "Warning, SCM Teardown"))
 	}
-}
-
-type PrepareNvmeRequest struct {
-	HugePageCount int
-	TargetUser    string
-	PCIWhitelist  string
-	ResetOnly     bool
-}
-
-// PrepareNvme preps locally attached SSDs and returns error.
-//
-// Suitable for commands invoked directly on server, not over gRPC.
-func (c *ControlService) PrepareNvme(req PrepareNvmeRequest) error {
-	ok, usr := common.CheckSudo()
-	if !ok {
-		return errors.Errorf("%s must be run as root or sudo", os.Args[0])
-	}
-
-	// falls back to sudoer or root if TargetUser is unspecified
-	tUsr := usr
-	if req.TargetUser != "" {
-		tUsr = req.TargetUser
-	}
-
-	// run reset first to ensure reallocation of hugepages
-	if err := c.nvme.spdk.reset(); err != nil {
-		return errors.WithMessage(err, "SPDK setup reset")
-	}
-
-	// if we're only resetting, just return before prep
-	if req.ResetOnly {
-		return nil
-	}
-
-	return errors.WithMessage(
-		c.nvme.spdk.prep(req.HugePageCount, tUsr, req.PCIWhitelist),
-		"SPDK setup",
-	)
-}
-
-type PrepareScmRequest struct {
-	Reset bool
-}
-
-// PrepareScm preps locally attached modules and returns need to reboot message,
-// list of pmem kernel devices and error directly.
-//
-// Suitable for commands invoked directly on server, not over gRPC.
-func (c *ControlService) PrepareScm(req PrepareScmRequest) (needsReboot bool, pmemDevs []pmemDev, err error) {
-	if err = c.scm.Setup(); err != nil {
-		err = errors.WithMessage(err, "SCM setup")
-		return
-	}
-
-	if !c.scm.initialized {
-		err = errors.New(msgScmNotInited)
-		return
-	}
-
-	if len(c.scm.modules) == 0 {
-		err = errors.New(msgScmNoModules)
-		return
-	}
-
-	if req.Reset {
-		// run reset to remove namespaces and clear regions
-		needsReboot, err = c.scm.PrepReset()
-		return
-	}
-
-	// transition to the next state in SCM preparation
-	return c.scm.Prep()
-}
-
-// ScanNvme scans locally attached SSDs and returns list directly.
-//
-// Suitable for commands invoked directly on server, not over gRPC.
-func (c *ControlService) ScanNvme() (types.NvmeControllers, error) {
-	if err := c.nvme.Discover(); err != nil {
-		return nil, errors.Wrap(err, "NVMe storage scan")
-	}
-
-	return c.nvme.controllers, nil
-}
-
-// ScanScm scans locally attached modules and returns list directly.
-//
-// Suitable for commands invoked directly on server, not over gRPC.
-func (c *ControlService) ScanScm() (types.ScmModules, error) {
-	if err := c.scm.Discover(); err != nil {
-		return nil, errors.Wrap(err, "SCM storage scan")
-	}
-
-	return c.scm.modules, nil
-}
-
-func (c *ControlService) StorageScan() ([]string, error) {
-	scanErrors := make([]error, 0, 2)
-	outStrs := make([]string, 0, 3)
-
-	outStrs = append(outStrs, "Scanning locally-attached storage...")
-
-	controllers, err := c.ScanNVMe()
-	if err != nil {
-		scanErrors = append(scanErrors, err)
-	} else {
-		outStrs = append(outStrs,
-			fmt.Sprintf("NVMe SSD controller and constituent namespaces:\n%s", controllers))
-	}
-
-	modules, err := c.ScanSCM()
-	if err != nil {
-		scanErrors = append(scanErrors, err)
-	} else {
-		outStrs = append(outStrs, fmt.Sprintf("SCM modules:\n%s", modules))
-	}
-
-	if len(scanErrors) > 0 {
-		errStr := "scan error(s):\n"
-		for _, err := range scanErrors {
-			errStr += fmt.Sprintf("  %s\n", err.Error())
-		}
-		return outStrs, errors.New(errStr)
-	}
-
-	return outStrs, nil
 }
 
 // awaitStorageFormat checks if running as root and server superblocks exist,
